+++ src/PHPUnit/Traits/AssertTraits.php
@@ -5,4 +5,5 @@ namespace VanCodX\Testing\PHPUnit\Traits;
 trait AssertTraits
 {
     use ExpectExceptionOnCallTrait;
+    use ExpectExceptionIfNotTrait;
 }