+++ src/PHPUnit/Traits/AssertTraits.php
@@ -4,5 +4,6 @@ namespace VanCodX\Testing\PHPUnit\Traits;
 
 trait AssertTraits
 {
-    use ExpectExceptionTrait;
+    use ExpectExceptionOnCallTrait;
+    use ExpectExceptionIfNotTrait;
 }