+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Change Log
 ==========
 
+1.7.0
+-----
+
+### Added:
+
+* ExpectExceptionIfNotTrait assertion method set.
+
+### Modified:
+
+* Renamed: ExpectExceptionTrait -> ExpectExceptionOnCallTrait.
+
 1.6.0
 -----